cleaning up test files
