+++ src/content/src/Tests/WebAPIService.Performance.Tests/packages.config
@@ -2,7 +2,6 @@
 <packages>
   <package id="Microsoft.Azure.KeyVault" version="3.0.1" targetFramework="net472" />
   <package id="Microsoft.Azure.KeyVault.WebKey" version="3.0.1" targetFramework="net472" />
-  <package id="Microsoft.Azure.Services.AppAuthentication" version="1.0.3" targetFramework="net472" />
   <package id="Microsoft.Extensions.Configuration" version="2.1.1" targetFramework="net472" />
   <package id="Microsoft.Extensions.Configuration.Abstractions" version="2.1.1" targetFramework="net472" />
   <package id="Microsoft.Extensions.Configuration.EnvironmentVariables" version="2.1.1" targetFramework="net472" />
@@ -12,7 +11,6 @@
   <package id="Microsoft.Extensions.FileProviders.Physical" version="2.1.1" targetFramework="net472" />
   <package id="Microsoft.Extensions.FileSystemGlobbing" version="2.1.1" targetFramework="net472" />
   <package id="Microsoft.Extensions.Primitives" version="2.1.1" targetFramework="net472" />
-  <package id="Microsoft.IdentityModel.Clients.ActiveDirectory" version="4.3.0" targetFramework="net472" />
   <package id="Microsoft.Rest.ClientRuntime" version="2.3.17" targetFramework="net472" />
   <package id="Microsoft.Rest.ClientRuntime.Azure" version="3.3.18" targetFramework="net472" />
   <package id="Newtonsoft.Json" version="11.0.2" targetFramework="net472" />
