--- src/content/src/Tests/WebAPIService.Unit.Tests/appsettings.json
@@ -1,2 +1,0 @@
-﻿{
-}
