+++ src/content/src/WebAPIService.sln
@@ -6,6 +6,7 @@ MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{5815314C-19FA-4746-8FE6-006713BC3072}"
 	ProjectSection(SolutionItems) = preProject
 		..\.gitignore = ..\.gitignore
+		AzureDevOps.testsettings = AzureDevOps.testsettings
 		codecoverage.runsettings = codecoverage.runsettings
 		Local.testsettings = Local.testsettings
 		nuget.config = nuget.config
